Split off nmea.CheckSyntax into separate file
